#issue-10-changing-code-following-the-comments
