+++ src/components/utils/format-date.tsx
@@ -1,0 +1,7 @@
+export default function FormatDate(dateString: string): string {
+  const actualDate = new Date(dateString);
+  const day = actualDate.getDate().toString().padStart(2, "0");
+  const month = (actualDate.getMonth() + 1).toString().padStart(2, "0");
+  const year = actualDate.getFullYear().toString();
+  return `${day}/${month}/${year}`;
+}
