+++ src/styles/globals.css
@@ -23,3 +23,15 @@
   background: #858c94;
   border-radius: 5px;
 }
+@layer components {
+  /* This won't be included in your compiled CSS unless you actually use it */
+  .header-component {
+    background: rgba(255, 255, 255, 0.2);
+    border-radius: 16px;
+    box-shadow: 0 4px 30px rgba(0, 0, 0, 0.1);
+    backdrop-filter: blur(5px);
+    -webkit-backdrop-filter: blur(5px);
+    border: 1px solid rgba(255, 255, 255, 0.3);
+  }
+}
+
